+++ runtime/src/Makefile.share
@@ -23,7 +23,8 @@ COMMON_LAUNCHER_SRCS = \
 	chpl-string-support.c \
 	chplcast.c \
 	config.c \
-	chpl-error.c
+	chpl-error.c \
+	chpl-unwind.c
 
 ONLY_LAUNCHER_SRCS = \
 	chpl-launcher-common.c
